[feat] Extract RISC-V SBI-ECALL function from l1_hypervisor/src/sbi.rs to arch/riscv/sbi.rs
