+++ arch/src/riscv/mod.rs
@@ -1,2 +1,3 @@
 pub mod cpu;
 pub mod instruction;
+pub mod sbi;
